Split out the download components
