+++ src/components/app/MenuButtons/Download.css
@@ -1,0 +1,12 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+.menuButtonsProfileDownloadPanel {
+  --offset-from-right: 25px;
+  --width: 30em;
+}
+
+.menuButtonsDownloadLink {
+  white-space: nowrap;
+}
